// src/utils/types.ts
export type FilterProps = {
  id: string;
  text: string;
};

export type RatingFilterProps = {
  id: string;
  img: HTMLImageElement;
  rating: number;
};

export type SidebarListProps = {
  id: string;
  icon: HTMLImageElement;
  text: string;
};

// export type actionFunction = {
//   prevState: any;
//   formData: FormData;
// }